refactor: order -> member 의존성 제거
